fix for crab pr test

+++ pr_testing/setup-pr-test-env.sh
@@ -67,7 +67,7 @@ if [ "${CMSSW_CVMFS_PATH}" != "" ] ; then
   popd
   mkdir -p $WORKSPACE/${CMSSW_VERSION}
   if [ -f ${CMSSW_CVMFS_PATH}/ibeos_cache.txt ] ; then ln -s ${CMSSW_CVMFS_PATH}/ibeos_cache.txt $WORKSPACE/${CMSSW_VERSION}/ibeos_cache.txt ; fi
-  ln -s ${CMSSW_CVMFS_PATH}/src $WORKSPACE/${CMSSW_VERSION}/src
+  [ -e $WORKSPACE/${CMSSW_VERSION}/src ] || ln -s ${CMSSW_CVMFS_PATH}/src $WORKSPACE/${CMSSW_VERSION}/src
 else
   pushd $WORKSPACE/$CMSSW_IB
     eval `scram run -sh` >/dev/null 2>&1
